Applied the same helper class structure as in PulseExtractor and PulseAnalyzer to predefined methods with PulseObjectGenerator and PredefinedGeneratorBase classes.
