+++ logic/predefined_generate_methods/basic_predefined_methods.py
@@ -21,9 +21,8 @@ top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi
 """
 
 import numpy as np
-import inspect
-from collections import OrderedDict
 from logic.pulse_objects import PulseBlockElement, PulseBlock, PulseBlockEnsemble, PulseSequence
+from logic.pulse_objects import PredefinedGeneratorBase
 from logic.sampling_functions import *
 
 
@@ -41,125 +40,12 @@ General Pulse Creation Procedure:
 """
 
 
-class PulsedObjectGenerator:
+class BasicPredefinedGenerator(PredefinedGeneratorBase):
     """
 
     """
-    def __init__(self, pulse_generator_settings, generation_parameters):
-        self._pulse_generator_settings = pulse_generator_settings
-        self._generation_parameters = generation_parameters
-        return
-
-    @classmethod
-    def attach_method(cls, method):
-        if callable(method) and (inspect.ismethod(method) or inspect.isfunction(method)):
-            # Bind the method as an attribute to cls
-            setattr(cls, method.__name__, method)
-        return
-
-    @property
-    def pulse_generator_settings(self):
-        return self._pulse_generator_settings
-
-    @pulse_generator_settings.setter
-    def pulse_generator_settings(self, settings_dict):
-        if isinstance(settings_dict, dict):
-            self._pulse_generator_settings = settings_dict
-        return
-
-    @property
-    def generation_parameters(self):
-        return self._generation_parameters
-
-    @generation_parameters.setter
-    def generation_parameters(self, settings_dict):
-        if isinstance(settings_dict, dict):
-            self._generation_parameters = settings_dict
-        return
-
-    @property
-    def predefined_generate_methods(self):
-        names = sorted([name for name in dir(self) if name.startswith('generate_')])
-        methods = OrderedDict()
-        for name in names:
-            methods[name[9:]] = getattr(self, name)
-        return methods
-
-    @property
-    def predefined_generate_params(self):
-        methods = self.predefined_generate_methods
-        for method_name, method in methods.items():
-            param_dict = dict()
-            signature_params = inspect.signature(method).parameters
-            for name, param in signature_params.items():
-                if param.default is param.empty:
-                    param_dict[name] = {'type': None, 'default': None}
-                else:
-                    param_dict[name] = {'type': type(param.default), 'default': param.default}
-            methods[method_name] = param_dict
-        return methods
-
-    @property
-    def channel_set(self):
-        channels = self._pulse_generator_settings.get('activation_config')
-        if channels is None:
-            channels = ('', set())
-        return channels[1]
-
-    @property
-    def analog_channels(self):
-        return {chnl for chnl in self.channel_set if chnl.startswith('a')}
-
-    @property
-    def digital_channels(self):
-        return {chnl for chnl in self.channel_set if chnl.startswith('d')}
-
-    @property
-    def laser_channel(self):
-        return self._generation_parameters.get('laser_channel')
-
-    @property
-    def sync_channel(self):
-        channel = self._generation_parameters.get('sync_channel')
-        return None if channel == '' else channel
-
-    @property
-    def gate_channel(self):
-        channel = self._generation_parameters.get('gate_channel')
-        return None if channel == '' else channel
-
-    @property
-    def analog_trigger_voltage(self):
-        return self._generation_parameters.get('analog_trigger_voltage')
-
-    @property
-    def laser_delay(self):
-        return self._generation_parameters.get('laser_delay')
-
-    @property
-    def microwave_channel(self):
-        channel = self._generation_parameters.get('microwave_channel')
-        return None if channel == '' else channel
-
-    @property
-    def microwave_frequency(self):
-        return self._generation_parameters.get('microwave_frequency')
-
-    @property
-    def microwave_amplitude(self):
-        return self._generation_parameters.get('microwave_amplitude')
-
-    @property
-    def laser_length(self):
-        return self._generation_parameters.get('laser_length')
-
-    @property
-    def wait_time(self):
-        return self._generation_parameters.get('wait_time')
-
-    @property
-    def rabi_period(self):
-        return self._generation_parameters.get('rabi_period')
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
     def generate_laser_on(self, name='laser_on', length=3.0e-6):
         """ Generates Laser on.
@@ -1029,226 +915,3 @@ class PulsedObjectGenerator:
         # append ensemble to created ensembles
         created_ensembles.append(block_ensemble)
         return created_blocks, created_ensembles, created_sequences
-
-    ################################################################################################
-    #                                   Helper methods                                          ####
-    ################################################################################################
-    def _get_idle_element(self, length, increment):
-        """
-        Creates an idle pulse PulseBlockElement
-
-        @param float length: idle duration in seconds
-        @param float increment: idle duration increment in seconds
-
-        @return: PulseBlockElement, the generated idle element
-        """
-        # Create idle element
-        return PulseBlockElement(init_length_s=length,
-                                 increment_s=increment,
-                                 pulse_function={chnl: Idle() for chnl in self.analog_channels},
-                                 digital_high={chnl: False for chnl in self.digital_channels})
-
-    def _get_trigger_element(self, length, increment, channels):
-        """
-        Creates a trigger PulseBlockElement
-
-        @param float length: trigger duration in seconds
-        @param float increment: trigger duration increment in seconds
-        @param str|list channels: The pulser channel(s) to be triggered.
-
-        @return: PulseBlockElement, the generated trigger element
-        """
-        if isinstance(channels, str):
-            channels = [channels]
-
-        # input params for element generation
-        pulse_function = {chnl: Idle() for chnl in self.analog_channels}
-        digital_high = {chnl: False for chnl in self.digital_channels}
-
-        # Determine analogue or digital trigger channel and set channels accordingly.
-        for channel in channels:
-            if channel.startswith('d'):
-                digital_high[channel] = True
-            else:
-                pulse_function[channel] = DC(voltage=self.analog_trigger_voltage)
-
-        # return trigger element
-        return PulseBlockElement(init_length_s=length,
-                                 increment_s=increment,
-                                 pulse_function=pulse_function,
-                                 digital_high=digital_high)
-
-    def _get_laser_element(self, length, increment):
-        """
-        Creates laser trigger PulseBlockElement
-
-        @param float length: laser pulse duration in seconds
-        @param float increment: laser pulse duration increment in seconds
-
-        @return: PulseBlockElement, two elements for laser and gate trigger (delay element)
-        """
-        return self._get_trigger_element(length=length,
-                                         increment=increment,
-                                         channels=self.laser_channel)
-
-    def _get_laser_gate_element(self, length, increment):
-        """
-        """
-        laser_gate_element = self._get_laser_element(length=length,
-                                                     increment=increment)
-        if self.gate_channel:
-            if self.gate_channel.startswith('d'):
-                laser_gate_element.digital_high[self.gate_channel] = True
-            else:
-                laser_gate_element.pulse_function[self.gate_channel] = DC(
-                    voltage=self.analog_trigger_voltage)
-        return laser_gate_element
-
-    def _get_delay_element(self):
-        """
-        Creates an idle element of length of the laser delay
-
-        @return PulseBlockElement: The delay element
-        """
-        return self._get_idle_element(length=self.laser_delay,
-                                      increment=0)
-
-    def _get_delay_gate_element(self):
-        """
-        Creates a gate trigger of length of the laser delay.
-        If no gate channel is specified will return a simple idle element.
-
-        @return PulseBlockElement: The delay element
-        """
-        if self.gate_channel:
-            return self._get_trigger_element(length=self.laser_delay,
-                                             increment=0,
-                                             channels=self.gate_channel)
-        else:
-            return self._get_delay_element()
-
-    def _get_sync_element(self):
-        """
-
-        """
-        return self._get_trigger_element(length=50e-9,
-                                         increment=0,
-                                         channels=self.sync_channel)
-
-    def _get_mw_element(self, length, increment, amp=None, freq=None, phase=None):
-        """
-        Creates a MW pulse PulseBlockElement
-
-        @param float length: MW pulse duration in seconds
-        @param float increment: MW pulse duration increment in seconds
-        @param float freq: MW frequency in case of analogue MW channel in Hz
-        @param float amp: MW amplitude in case of analogue MW channel in V
-        @param float phase: MW phase in case of analogue MW channel in deg
-
-        @return: PulseBlockElement, the generated MW element
-        """
-        if self.microwave_channel.startswith('d'):
-            mw_element = self._get_trigger_element(
-                length=length,
-                increment=increment,
-                channels=self.microwave_channel)
-        else:
-            mw_element = self._get_idle_element(
-                length=length,
-                increment=increment)
-            mw_element.pulse_function[self.microwave_channel] = Sin(amplitude=amp,
-                                                                    frequency=freq,
-                                                                    phase=phase)
-        return mw_element
-
-    def _get_multiple_mw_element(self, length, increment, amps=None, freqs=None, phases=None):
-        """
-        Creates single, double or triple sine mw element.
-
-        @param float length: MW pulse duration in seconds
-        @param float increment: MW pulse duration increment in seconds
-        @param amps: list containing the amplitudes
-        @param freqs: list containing the frequencies
-        @param phases: list containing the phases
-        @return: PulseBlockElement, the generated MW element
-        """
-        if isinstance(amps, (int, float)):
-            amps = [amps]
-        if isinstance(freqs, (int, float)):
-            freqs = [freqs]
-        if isinstance(phases, (int, float)):
-            phases = [phases]
-
-        if self.microwave_channel.startswith('d'):
-            mw_element = self._get_trigger_element(
-                length=length,
-                increment=increment,
-                channels=self.microwave_channel)
-        else:
-            mw_element = self._get_idle_element(
-                length=length,
-                increment=increment)
-
-            sine_number = min(len(amps), len(freqs), len(phases))
-
-            if sine_number < 2:
-                mw_element.pulse_function[self.microwave_channel] = Sin(amplitude=amps[0],
-                                                                        frequency=freqs[0],
-                                                                        phase=phases[0])
-            elif sine_number == 2:
-                mw_element.pulse_function[self.microwave_channel] = DoubleSin(amplitude_1=amps[0],
-                                                                              amplitude_2=amps[1],
-                                                                              frequency_1=freqs[0],
-                                                                              frequency_2=freqs[1],
-                                                                              phase_1=phases[0],
-                                                                              phase_2=phases[1])
-            else:
-                mw_element.pulse_function[self.microwave_channel] = TripleSin(amplitude_1=amps[0],
-                                                                              amplitude_2=amps[1],
-                                                                              amplitude_3=amps[2],
-                                                                              frequency_1=freqs[0],
-                                                                              frequency_2=freqs[1],
-                                                                              frequency_3=freqs[2],
-                                                                              phase_1=phases[0],
-                                                                              phase_2=phases[1],
-                                                                              phase_3=phases[2])
-        return mw_element
-
-    def _get_mw_laser_element(self, length, increment, amp=None, freq=None, phase=None):
-        """
-
-        @param length:
-        @param increment:
-        @param amp:
-        @param freq:
-        @param phase:
-        @return:
-        """
-        mw_laser_element = self._get_mw_element(length=length,
-                                                increment=increment,
-                                                amp=amp,
-                                                freq=freq,
-                                                phase=phase)
-        if self.laser_channel.startswith('d'):
-            mw_laser_element.digital_high[self.laser_channel] = True
-        else:
-            mw_laser_element.pulse_function[self.laser_channel] = DC(
-                voltage=self.analog_trigger_voltage)
-        return mw_laser_element
-
-    def _get_ensemble_count_length(self, ensemble, created_blocks):
-        """
-
-        @param ensemble:
-        @param created_blocks:
-        @return:
-        """
-        if self.gate_channel:
-            length = self.laser_length + self.laser_delay
-        else:
-            blocks = {block.name: block for block in created_blocks}
-            length = 0.0
-            for block_name, reps in ensemble.block_list:
-                length += blocks[block_name].init_length_s * (reps + 1)
-                length += blocks[block_name].increment_s * ((reps**2 + reps) / 2)
-        return length
